requirements.txt
asgiref==3.11.0
Django==5.2.10
djangorestframework==3.16.1
drf-yasg==1.21.14
gunicorn==25.1.0
inflection==0.5.1
packaging==26.0
psycopg2-binary==2.9.11
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
